chore: extract get to core, add withPlaiceholder
