+++ examples/next/package.json
@@ -17,7 +17,6 @@
     "@plaiceholder/base64": "^1.0.0",
     "@plaiceholder/blurhash": "^1.0.0",
     "@plaiceholder/css": "^1.0.0",
-    "@plaiceholder/get": "^1.0.0",
     "@plaiceholder/svg": "^1.0.0",
     "@plaiceholder/tailwindcss": "^1.0.0",
     "@tailwindcss/aspect-ratio": "0.2.0",
@@ -26,8 +25,7 @@
     "react": "17.0.2",
     "react-blurhash": "0.1.3",
     "react-dom": "17.0.2",
-    "sharp": "0.28.1",
-    "image-size": "0.9.7"
+    "sharp": "0.28.1"
   },
   "devDependencies": {
     "autoprefixer": "10.2.5",
