+++ packages/core/package.json
@@ -33,5 +33,8 @@
   "homepage": "https://github.com/joe-bell/plaiceholder#readme",
   "peerDependencies": {
     "sharp": ">= 0.26.2"
+  },
+  "dependencies": {
+    "image-size": "0.9.7"
   }
 }
