+++ packages/core/src/index.ts
@@ -1,4 +1,9 @@
+import fs from "fs";
+import path from "path";
 import sharp from "sharp";
+import { promisify } from "util";
+import sizeOf from "image-size";
+import type { ISizeCalculationResult } from "image-size/dist/types/interface";
 
 export const ACCEPTED_FILE_TYPES = ["jpeg", "png"] as const;
 
@@ -49,3 +54,36 @@ export const getPixels: GetPixels = async (imageBuffer) =>
         if (err) return reject(err);
       });
   });
+
+export type GetImagePath = string;
+
+export interface GetImage {
+  (imagePath: GetImagePath): Promise<
+    {
+      buffer: PlaiceholderImage;
+    } & ISizeCalculationResult
+  >;
+}
+
+export const getImage: GetImage = async (imagePath) => {
+  if (imagePath.startsWith("http")) {
+    throw new Error("Sorry, remote images aren't supported just yet");
+  }
+
+  if (!imagePath.startsWith("/")) {
+    throw new Error(
+      `Failed to parse \"${imagePath}\", relative images must start with a leading slash`
+    );
+  }
+
+  const buffer = await promisify(fs.readFile)(
+    path.join("./public/", imagePath)
+  );
+
+  const details = sizeOf(buffer);
+
+  return {
+    buffer,
+    ...details,
+  };
+};
