--- packages/get/README.md
@@ -1,1 +1,0 @@
-[See main README](https://github.com/joe-bell/plaiceholder)
